backend/requirements.txt
fastapi
sentence_transformers
pydantic
google-generativeai
uvicorn
python-dotenv
redis
